requirements.txt
streamlit>=1.30.0
pypdf
google-genai
tavily-python